Processing of complete image added
Actually not running - never did since Max did stuff :)

+++ Source/functions/AbstractionLayers/Layer_SURFFeatures/AbstractionLayer_SURFFeatures_Properties.h
@@ -9,7 +9,7 @@ public:
     AbstractionLayer_SURFFeatures_Properties() {}
 
 private:
-    uint16_t m_numberOfFeaturesDetected;
+    float m_numberOfFeaturesDetected;
     friend class AbstractionLayer_SURFFeatures;
 };
